+++ packages/polythene-core/src/developer.js
@@ -1,6 +1,16 @@
 
+const warn = (old, replacement) =>
+  console.warn(`${old} is deprecated and will be removed in a later version of Polythene. Use '${replacement}' instead.`); // eslint-disable-line no-console
+
 export const deprecation = (component, { option, newOption, newComponent }) => (
-  option && console.warn(`${component}: option '${option}' is deprecated and will be removed in later versions. Use '${newOption}' instead.`), // eslint-disable-line no-console
-  newComponent && !newOption && console.warn(`${component}: this component is deprecated and will be removed in later versions. Use '${newComponent}' instead.`), // eslint-disable-line no-console
-  newComponent && newOption && console.warn(`${component}: this component is deprecated and will be removed in later versions. Use '${newComponent}' with option '${newOption}' instead.`) // eslint-disable-line no-console
+  option && warn(`${component}: option '${option}'`, newOption),
+  newComponent && !newOption && warn(`${component}: this component`, newComponent),
+  newComponent && newOption && warn(`${component}: this component`, `'${newComponent}' with option '${newOption}'`)
 );
+
+export const deprecationForElementAttrs = (component, { attrs, deprecated, keys }) => {
+  deprecated.forEach(attr => {
+    const attribute = keys[attr] || attr;
+    attrs[attribute] !== undefined && deprecation(component, { option: attribute, newOption: `elementAttrs: { ${[attribute]}: ${attrs[attribute]} }` });
+  });
+};