+++ packages/polythene-core/src/index.js
@@ -7,5 +7,5 @@ export { Multi } from "./multi";
 export { show, hide, transitionComponent } from "./transition";
 export { throttle, subscribe, unsubscribe, emit } from "./events";
 export { getStyle, stylePropCompare, isRTL, styleDurationToMs } from "./style";
-export { deprecation } from "./developer";
+export { deprecation, deprecationForElementAttrs } from "./developer";
 export { iconDropdownUp, iconDropdownDown } from "./icons";